Use a single build script and pass in the NODE_ENV
