--- cli/testdata/checkout-ui-extension/production.js
@@ -1,3 +1,0 @@
-const build = require('./build');
-
-build({mode: 'production'})
